Add check for Qualcomm devices on Windows
Some of these devices are not capable of running >=4.0, due to
issues with Mesa's Compute Shaders and their D3D drivers.

This PR marks those GPUs as unsupported, and prints info to stdout.

A driver update will be available for 8cx Gen3 and newer.

This issue applies even when using emulated x64.