Make the bNodeTree pointer in enum sockets into a truly weak reference.
No refcounting is needed for this pointer, we do not want to keep the
node tree alive through an enum socket reference. Currently the tree
is always referenced by an internal node group that the socket is part
of, but even without that the enum reference should be allowed to become
invalid.