Fix performance regression on Metal/AMD due to new BSDFs
The increase amount of BSDF code from Principled BSDF v2 and the
microfacet BSDF led to a big performance regression on Metal and AMD.
We have not been able to find a good workaround for all scenes.

This change disables the Principled Hair BSDF code when it is not used
in the scene. This makes such scenes faster, but performance is still
bad in scenes that do use it.

Ref #112596