UI: let keymap header take more horizontal space when filtering
The UI displaying keymaps in the User Preferences is different
depending on whether the filter is being used or not. If it is used,
the header for each keymap is cut short using the default window size.
Even resizing the window has little effect. For instance, searching
for "grease" will display:

- Grease Pencil ...
- Grease Pencil ...
- Grease Pencil ...

instead of:

- Grease Pencil Stroke Edit Mode
- Grease Pencil Stroke Paint Mode
- Grease Pencil Stroke Paint (Draw brush)

This commit changes the UI so that the title takes up all available
space. It reuses the code from the case where the filter is empty.
This makes the Restore button narrower, matching the no-filter
version.