GPv3: Flatten grease pencil layers for the conversion nodes
This changes the following nodes:
* Convex Hull
* Curve to Mesh
* Curve to Points
* Fill Curve
* Instance on Points

This is a rather naif approach, but rather quick to put together.
Basically I'm realizing the instances for the nodes that were creating
instances before.

For the nodes that don't propagate info (convex hull, curve fill) no
layer attribute is being propagated.

The instance on points is not propagating layer attributes either.