UI Code Quality: Rename `UI_ACTIVE` button flag to `UI_HOVER_FOCUS`
Calling this state "Active" has been confusing for a long time, it's not
clear that this is essentially a mouse hover state. It's especially easy
to confuse this with the "Select" state (`UI_SELECT` button flag), which
is used to indicate a pushed button state (and should be renamed too).

The term "Hover Focus" was chosen because it both reflects how it
indicates the hovered state, but also the fact that such buttons can
receive input like shortcut events. A comment is added explaining this
state flag.

See #112160.