Fix #112293: wrong console line cursor position with non-ASCII chars
The ConsoleLine's `cursor` stores the index of the char where the
cursor is currently at. This works fine for ASCII characters, which
are all one-byte long, but will be offset when the string contains
non-ASCII chars.

This caused an issue during autocomplete where the matching results
would be offset to the right.

This commit reuses the same logic as the text editor's
`current_character` RNA property, so that on getting and setting, the
cursor index is converted from and to UTF-8.