GPv3: Copy/Paste Keyframes
This commit implements Copy and Past Keyframes for GPv3. The operation
is available in Animation workspace. The copy operator copies the
selected GPv3 Keyframes and the paste operator paste the Keyframe from
the clipboard.

**Technical notes**
`grease_pencil_copy_keyframes()` iterates through every selected frame
for every channel/layer, then stores frame_number and layer name in a
vector of `BufferItem`. Please note that layer can be moved so storing
layer pointer is not viable.
`grease_pencil_paste_keyframes()` iterates through all `BufferItem` and
uses `GreasePencil::insert_duplicate_frame()` to duplicate the keyframe
and drawing. Please note that there's a issue with the function that may
lead to crash, for which I opened #114352.