GPv3: Rename `stroke_color_set` to `stroke_material_set`
The old name was logic when the color of the stroke was using a
palette and not materials, but now always a material is used
so the name "color" is not logic.

As we are changing all the API, now is the moment to set the right name.

Also, the commit includes some cleanup of the code.