Realtime Compositor: Improve classic Kuwahara precision
This patch improves the precision of the classic Kuwahara operation by
allowing more precise sum evaluation of summed area tables.

The improved precision is achieved by delaying the block completion pass
of the algorithm which implements equation (41) to evaluation time, and
separating the SAT evaluation into an inter-block prologue sum values
and intra-block values. The inter-block values are typically very large,
while the intra-block values are much smaller, so the idea is to compute
the SAT sum on the inter-block and intra-block values separately, then
sum them at the end.

This is of course more expensive and more memory intensive to compute
due to duplicate work at evaluation time and more textures fetches.

While the improved precision is measurably and visually better, the
results are still not really usable due to the artifacts. So this is
more of an infeasibility proof of concept rather than an actual patch.