VSE: remove speed_factor RNA property from sound strips
Property `speed_factor` was used before retiming was implemented and
kept for potential versioning code for complex speed animation, that was
possible with even older `pitch` property. Files using static values
are already converted.

Such versioning code would be quite complex, possibly slow and maybe
could corrupt files. This is due to multiple factors:
- Sound seeking was broken, so conversion would have to ignore all
  keyframes before strip starts to map frames properly.
- Because some animation was effectively ignored, it may cause
  inconsistencies when doing conversion.
- It would have to integrate value of `speed_factor` to map keyframes
  to strip space, but the animation is not limited to strip length.
- For each keyframe where speed gradually changes, at least one smooth
  speed transition would be required, but there would be discrepancies
  that would have to be accounted for. With simpler strategy it is
  likely, that extent of ramps would be limited and thus animation would
  be quite different from original.

So because of these reasons, I think it is best to not convert
`speed_factor` animation.