UI: Snap Menu - Match the symbols we use in the viewport
This changes the icons we use in the Snap menu to match the symbols we
use while snapping. Note that we use the same icon for Face and Volume,
which means in the UI we can skip showing the icon for Volume.

Snap symbols introduced on fb556c75df4c.