PyAPI: Speed up Mesh.shade_flat() using foreach_set
Replace the slower iteration in Python with using foreach_set and a
compatible buffer object filled with `True`.

For meshes under 24 faces, the new code is slightly slower due to higher
overhead, but the increase in performance as the number of faces
increases should more than make up for this.

NumPy is only used to create a buffer filled with `True` that is
compatible with `foreach_set`. The `array` module does not support the
`bool` C type, and `memoryview` objects and ctypes arrays are not
currently considered compatible buffers by `foreach_set`.

No changes, other than `Mesh.shade_flat()` performance, are expected
with this patch.

Time per call for me for meshes with:
- 24 faces:
  - Before: ~0.0055ms
  - After: ~0.0050ms
  - ~1.1 times faster
- 1536 faces:
  - Before: ~0.24ms
  - After: ~0.038ms
  - ~6.2 times faster
- 24576 faces:
  - Before: ~3.83ms
  - After: ~0.54ms
  - ~7.1 times faster
- 393216 faces:
  - Before: ~59.5ms
  - After: ~8.6ms
  - ~6.9 times faster