Fix #114414: Drawing Artifacts Single Channel JPEG Files.
Single channel JPEG files are stored internally as 4 channels, where
RGB shader the same value. A is fully opaque. The number of bit planes
of this image format is still set to 8. When creating the GPU texture
the number of bitplanes is used to identfy if the data could be stored
on the GPU as a single channel image. But in this case the data loaded
in the GPU texture would contain 4 channels.

This PR fixes the issue to also check the number of channels. Although
for rendering it would be better that the Image buffer would only
contain a single channel, it would require more testing