Fix various crashes caused by static multifunctions in Menu Switch node.
Unlike most other node multifunctions, the menu switch function depends
on the concrete enum definition. The input value is compared to the item
identifier values, which are generally different for each instance of
the node. So the multifunction and its signature have to be constructed
on a case-by-case basis and cannot be constructed in advance.