Fix #114609: Don't emit text for compose sequence continuations
Previously, for any compose sequence without dead keys (for example,
<Multi_key> <a> <grave> resulting in "à") wouldn't ignore the grave,
as it's not a dead key. The result of typing this compose sequence
in a textbox would contain both characters: "aà".

To fix this, I've modified the switch statement in
`xkb_compose_state_feed_and_get_utf8` to consider a status of
`...COMPOSING` to be handled.