Build: prevent OpenSSL using /etc/ssl on macOS & Linux
Using /etc/ssl only makes sense when the versions of SSL on the system
is compatible with the version Blender uses.

Failure to load the configuration for e.g. causes SSL to fail entirely
(causing downloading over HTTPS to fail).

Recently [0] de facto standard directory `/etc/ssl` was used however
this can back-fire, see: #114452 for details.

Instead of reverting the define, point to a location known not to exist.

[0]: 60a8ae7830acd09c3d8e9ab52630d48ecd4d281b