Fix incorrect behavior and bad memory access introduced with the refactor
At least one of the issues was the following code in SCULPT_search_sphere():
  center = data->ss->cache ? data->ss->cache->location : data->ss->cursor_location;

The `location` and `cursor_location` had different type, which is not how the ternary operator could work in this case, so the cursor_location got implicitly converted from float[3] to float3, and then to a pointer.
This lead to center being assigned to a pointer of a temporary object.

Not sure how to fully guard about such error automatically. Solution is to convert even more float[3] to float3.