GPU:Refactor:attachment_set
One change that was required due to the refactoring of render passes is WIP.
The reason for this change is that when we configure a render pass,
we need to go through initiation and finalization processes.
but, i would like to consider the state of `Subpass-API` and consider its consistency.