GPU:Refactor:FrameBuffer configuration
When configuring Framebuffer, Vulkan allows you to configure VkRenderpass.
This is because Blender's concept of framebuffer includes Vulkan's concept of render passes.
In Vulkan, a VkFramebuffer is an Object that is directly associated with a resource and is handled in a narrower category.
Therefore, it is desirable to make the function that configures the frame buffer a virtual function.