Fix #114536: Redo panel to small when opening 4.0 files in 3.6
Forward compatibility fix for the 3.6 branch only.

Problem is 4.0 sets the RGN_FLAG_NO_USER_RESIZE in versioning for
a number of regions. When reading 4.0 files this would match the
RGN_FLAG_PREFSIZE_OR_HIDDEN flag, causing wrong region sizes.

I experimented with only clearing the flag for the redo panel, but I
noticed the flag would also mess with the File Browser region showing
the file path when saved in a layout. So clear it for all regions that
don't intentionally use it.

Another option is to not reuse the flag-bit in 4.0. However reusing
flag-bits should be fine, such forward compatibility patches are not a
big deal IMO.